+++ wand/image.py
@@ -2387,10 +2387,14 @@ class Image(BaseImage):
                    imagemagick detect the file format. Used only in
                    ``blob`` or ``file`` cases
     :type format: :class:`basestring`
-    :param width: the width of new blank image.
+    :param width: the width of new blank image or an image loaded from raw
+                  data.
     :type width: :class:`numbers.Integral`
-    :param height: the height of new blank imgage.
+    :param height: the height of new blank imgage or an image loaded from
+                   raw data.
     :type height: :class:`numbers.Integral`
+    :param depth: the depth used when loading raw data.
+    :type depth: :class:`numbers.Integral`
     :param background: an optional background color.
                        default is transparent
     :type background: :class:`wand.color.Color`
@@ -2413,6 +2417,14 @@ class Image(BaseImage):
 
     .. versionadded:: 0.3.0
        The ``resolution`` parameter.
+
+    .. versionadded:: 0.4.2
+       The ``depth`` parameter.
+
+    .. versionchanged:: 0.4.2
+       The ``depth``, ``width`` and ``height`` parameters can be used
+       with the ``filename``, ``file`` and ``blob`` parameters to load
+       raw pixel data.
 
     .. describe:: [left:right, top:bottom]
 
@@ -2459,37 +2471,36 @@ class Image(BaseImage):
     channel_depths = None
 
     def __init__(self, image=None, blob=None, file=None, filename=None,
-                 format=None, width=None, height=None, background=None,
-                 resolution=None):
-        new_args = width, height, background
-        open_args = image, blob, file, filename
-        if (any(a is not None for a in new_args) and
-                any(a is not None for a in open_args)):
+                 format=None, width=None, height=None, depth=None,
+                 background=None, resolution=None):
+        new_args = width, height, background, depth
+        open_args = blob, file, filename
+        if any(a is not None for a in new_args) and image is not None:
             raise TypeError("blank image parameters can't be used with image "
-                            'opening parameters')
-        elif any(a is not None and b is not None
-                 for i, a in enumerate(open_args)
-                 for b in open_args[:i] + open_args[i + 1:]):
-            raise TypeError('parameters are exclusive each other; use only '
-                            'one at once')
-        elif not (format is None or isinstance(format, string_type)):
-            raise TypeError('format must be a string, not ' + repr(format))
+                            'parameter')
+        if sum(a is not None for a in open_args + (image,)) > 1:
+            raise TypeError(', '.join(open_args) +
+                            ' and image parameters are exclusive each other; '
+                            'use only one at once')
+        if not (format is None):
+            if not isinstance(format, string_type):
+                raise TypeError('format must be a string, not ' + repr(format))
+            if not any(a is not None for a in open_args):
+                raise TypeError('format can only be used with the blob, file '
+                                'or filename parameter')
+        if depth not in [None, 8, 16, 32]:
+            raise ValueError('Depth must be 8, 16 or 32')
         with self.allocate():
             if image is None:
                 wand = library.NewMagickWand()
                 super(Image, self).__init__(wand)
-            if width is not None and height is not None:
-                self.blank(width, height, background)
-            elif image is not None:
+            if image is not None:
                 if not isinstance(image, BaseImage):
                     raise TypeError('image must be a wand.image.Image '
                                     'instance, not ' + repr(image))
-                elif format:
-                    raise TypeError('format option cannot be used with image '
-                                    'nor filename')
                 wand = library.CloneMagickWand(image.wand)
                 super(Image, self).__init__(wand)
-            else:
+            elif any(a is not None for a in open_args):
                 if format:
                     format = binary(format)
                 with Color('transparent') as bg:  # FIXME: parameterize this
@@ -2497,23 +2508,40 @@ class Image(BaseImage):
                                                               bg.resource)
                     if not result:
                         self.raise_exception()
+
+                # allow setting the width, height and depth
+                # (needed for loading raw data)
+                if width is not None and height is not None:
+                    if not isinstance(width, numbers.Integral) or width < 1:
+                        raise TypeError('width must be a natural number, '
+                                        'not ' + repr(width))
+                    if not isinstance(height, numbers.Integral) or height < 1:
+                        raise TypeError('height must be a natural number, '
+                                        'not ' + repr(height))
+                    library.MagickSetSize(self.wand, width, height)
+                if depth is not None:
+                    library.MagickSetDepth(self.wand, depth)
+                if format:
+                    library.MagickSetFormat(self.wand, format)
+                    if not filename:
+                        library.MagickSetFilename(self.wand,
+                                                  b'buffer.' + format)
                 if file is not None:
-                    if format:
-                        library.MagickSetFilename(self.wand,
-                                                  b'buffer.' + format)
                     self.read(file=file, resolution=resolution)
-                if blob is not None:
-                    if format:
-                        library.MagickSetFilename(self.wand,
-                                                  b'buffer.' + format)
+                elif blob is not None:
                     self.read(blob=blob, resolution=resolution)
                 elif filename is not None:
-                    if format:
-                        raise TypeError(
-                            'format option cannot be used with image '
-                            'nor filename'
-                        )
                     self.read(filename=filename, resolution=resolution)
+                # clear the wand format, otherwise any subsequent call to
+                # MagickGetImageBlob will silently change the image to this
+                # format again.
+                library.MagickSetFormat(self.wand, binary(""))
+            elif width is not None and height is not None:
+                self.blank(width, height, background)
+                if depth:
+                    r = library.MagickSetImageDepth(self.wand, depth)
+                    if not r:
+                        raise self.raise_exception()
             self.metadata = Metadata(self)
             from .sequence import Sequence
             self.sequence = Sequence(self)
@@ -2714,7 +2742,7 @@ class Image(BaseImage):
         .. versionadded:: 0.3.6
 
         """
-        compression_index = libmagick.MagickGetImageCompression(self.wand)
+        compression_index = library.MagickGetImageCompression(self.wand)
         return COMPRESSION_TYPES[compression_index]
 
     @compression.setter